+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ship-mate-hidden-recompile` now prompts after customizable
   `ship-mate-prompt-for-hidden-buffer-idle-delay`.
 - Command `ship-mate-edit-history` to edit the history.
+- A lighter that provides the `ship-mate-command-map` as a menu.
 
 ### Changed